feat: Move pin code to `gpio` module
