+++ src/swm.rs
@@ -6,14 +6,12 @@
 use lpc82x;
 use lpc82x::swm::pinenable0;
 
-use ::{
-    syscon,
-    Pin,
-};
+use gpio::Pin;
 use init_state::{
     self,
     InitState,
 };
+use syscon;
 
 
 /// Interface to the switch matrix (SWM)
@@ -98,35 +96,35 @@ macro_rules! impl_pin_ext {
     }
 }
 
-impl_pin_ext!(::PIO0_0<'gpio> , ACMP_I1);
-impl_pin_ext!(::PIO0_1<'gpio> , ACMP_I2, CLKIN);
-impl_pin_ext!(::PIO0_2<'gpio> , SWDIO);
-impl_pin_ext!(::PIO0_3<'gpio> , SWCLK);
-impl_pin_ext!(::PIO0_4<'gpio> , ADC_11);
-impl_pin_ext!(::PIO0_5<'gpio> , RESETN);
-impl_pin_ext!(::PIO0_6<'gpio> , VDDCMP, ADC_1);
-impl_pin_ext!(::PIO0_7<'gpio> , ADC_0);
-impl_pin_ext!(::PIO0_8<'gpio> , XTALIN);
-impl_pin_ext!(::PIO0_9<'gpio> , XTALOUT);
-impl_pin_ext!(::PIO0_10<'gpio>, I2C0_SCL);
-impl_pin_ext!(::PIO0_11<'gpio>, I2C0_SDA);
-impl_pin_ext!(::PIO0_12<'gpio>);
-impl_pin_ext!(::PIO0_13<'gpio>, ADC_10);
-impl_pin_ext!(::PIO0_14<'gpio>, ACMP_I3, ADC_2);
-impl_pin_ext!(::PIO0_15<'gpio>);
-impl_pin_ext!(::PIO0_16<'gpio>);
-impl_pin_ext!(::PIO0_17<'gpio>, ADC_9);
-impl_pin_ext!(::PIO0_18<'gpio>, ADC_8);
-impl_pin_ext!(::PIO0_19<'gpio>, ADC_7);
-impl_pin_ext!(::PIO0_20<'gpio>, ADC_6);
-impl_pin_ext!(::PIO0_21<'gpio>, ADC_5);
-impl_pin_ext!(::PIO0_22<'gpio>, ADC_4);
-impl_pin_ext!(::PIO0_23<'gpio>, ACMP_I4, ADC_3);
-impl_pin_ext!(::PIO0_24<'gpio>);
-impl_pin_ext!(::PIO0_25<'gpio>);
-impl_pin_ext!(::PIO0_26<'gpio>);
-impl_pin_ext!(::PIO0_27<'gpio>);
-impl_pin_ext!(::PIO0_28<'gpio>);
+impl_pin_ext!(::gpio::PIO0_0<'gpio> , ACMP_I1);
+impl_pin_ext!(::gpio::PIO0_1<'gpio> , ACMP_I2, CLKIN);
+impl_pin_ext!(::gpio::PIO0_2<'gpio> , SWDIO);
+impl_pin_ext!(::gpio::PIO0_3<'gpio> , SWCLK);
+impl_pin_ext!(::gpio::PIO0_4<'gpio> , ADC_11);
+impl_pin_ext!(::gpio::PIO0_5<'gpio> , RESETN);
+impl_pin_ext!(::gpio::PIO0_6<'gpio> , VDDCMP, ADC_1);
+impl_pin_ext!(::gpio::PIO0_7<'gpio> , ADC_0);
+impl_pin_ext!(::gpio::PIO0_8<'gpio> , XTALIN);
+impl_pin_ext!(::gpio::PIO0_9<'gpio> , XTALOUT);
+impl_pin_ext!(::gpio::PIO0_10<'gpio>, I2C0_SCL);
+impl_pin_ext!(::gpio::PIO0_11<'gpio>, I2C0_SDA);
+impl_pin_ext!(::gpio::PIO0_12<'gpio>);
+impl_pin_ext!(::gpio::PIO0_13<'gpio>, ADC_10);
+impl_pin_ext!(::gpio::PIO0_14<'gpio>, ACMP_I3, ADC_2);
+impl_pin_ext!(::gpio::PIO0_15<'gpio>);
+impl_pin_ext!(::gpio::PIO0_16<'gpio>);
+impl_pin_ext!(::gpio::PIO0_17<'gpio>, ADC_9);
+impl_pin_ext!(::gpio::PIO0_18<'gpio>, ADC_8);
+impl_pin_ext!(::gpio::PIO0_19<'gpio>, ADC_7);
+impl_pin_ext!(::gpio::PIO0_20<'gpio>, ADC_6);
+impl_pin_ext!(::gpio::PIO0_21<'gpio>, ADC_5);
+impl_pin_ext!(::gpio::PIO0_22<'gpio>, ADC_4);
+impl_pin_ext!(::gpio::PIO0_23<'gpio>, ACMP_I4, ADC_3);
+impl_pin_ext!(::gpio::PIO0_24<'gpio>);
+impl_pin_ext!(::gpio::PIO0_25<'gpio>);
+impl_pin_ext!(::gpio::PIO0_26<'gpio>);
+impl_pin_ext!(::gpio::PIO0_27<'gpio>);
+impl_pin_ext!(::gpio::PIO0_28<'gpio>);
 
 
 /// Implemented for types that represent movable functions
